requirements.txt
google-genai
flask
qrcode[pil]
pillow
flask-sqlalchemy
flask-migrate